use overwriteClassOnLicense
