--- apps/meteor/ee/app/contact-id-verification/server/hooks/index.ts
@@ -1,1 +1,0 @@
-import './verifyContactChannel.ts';
