+++ apps/meteor/ee/app/contact-id-verification/server/index.ts
@@ -1,5 +1,1 @@
-import { License } from '@rocket.chat/license';
-
-await License.onLicense('chat.rocket.contact-id-verification', async () => {
-	await import('./hooks');
-});
+import './services/omnichannelService';
